Fixed tooltip on images in gallery.
The title of the anchor tag was been used for the caption in Slimbox. Since the
description was wanted for the caption, it was used in the anchor tag, which
was then displayed as the tooltip as well.

Fortuantly, Slimbox2 has a nice hook to specify the caption anyway wanted,
so the alt text on the first child image of the anchor is used for the caption
instead.

Closes #20